services/billing_admin/pyproject.toml
[tool.poetry]
name = "billing_admin"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.1"
django-model-utils = "4.0.0"
psycopg2-binary = "2.8.6"
django-postgres-extra = "2.0.2"
celery = "5.1.0"
django-celery-beat = "2.2.0"
django-celery-results = "2.1.0"
requests = "^2.26.0"
boto3 = "^1.18.9"
pycurl = "^7.43.0"
yookassa = "^2.1.2"
stripe = "^2.60.0"
gunicorn = "20.0.4"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "2.2"
django-extensions = "3.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
